Add files via upload
# Cat App

The Cat App is a beginner-friendly web project that showcases various popular cat breeds using basic HTML tags. It provides information about each cat breed, including their names, characteristics, and images. This project is ideal for those who are new to web development and want to learn the fundamentals of HTML. It was created as part of the FreeCodeCamp curriculum.